pyproject.toml
[project]
name = "grasp_agents"
version = "0.7.0"
description = "Grasp Agents Library"
readme = "README.md"
requires-python = ">=3.11.4,<4"
dependencies = [
    "httpx>=0.27.0,<1",
    "openai==2.23.0",
    "tenacity>=8.3.0,<9",
    "tqdm>=4.66.2,<5",
    "pyyaml>=6.0.2",
    "pydantic>=2",
    "litellm==1.81.13",
    "anthropic==0.84.0",
    "opentelemetry-sdk>=1.37.0",
    "traceloop-sdk>=0.47.2",
    "termcolor>=2.4.0,<3",
    "dotenv>=0.9.9",
    "google-genai>=1.65.0",
    "numpy<2",
]

[dependency-groups]
dev = [
    "ipython>=8.3.0,<9",
    "ipykernel>=6.19.1,<7",
    "ipywidgets>=8.0.4,<9",
    "widgetsnbextension>=4.0.5,<5",
    "types-cachetools>=5.0.1,<6",
    "pre-commit-uv>=4.1.4",
    "twine>=5.1.1,<6",
    "ruff>=0.11.8",
    "pyright>=1.1.400",
    "pip-licenses>=5.0.0",
    "pytest>=8,<9",
    "pytest-asyncio>=1.3.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
observability = [
    "arize-phoenix-otel>=0.13.1",
    "openinference-instrumentation-openllmetry>=0.1.2",
    "openinference-instrumentation-openai>=0.1.32",
    "openinference-instrumentation-litellm>=0.1.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "strict"
markers = ["integration: requires real API keys (skipped by default)"]
addopts = "-m 'not integration'"

[tool.hatch.build.targets.sdist]
include = ["src/grasp_agents", "src/grasp_agents_kits"]
exclude = ["src/grasp_agents/examples", "*.ipynb"]

[tool.hatch.build.targets.wheel]
packages = ["src/grasp_agents", "src/grasp_agents_kits"]
exclude = ["src/grasp_agents/examples", "*.ipynb"]

[tool.hatch.metadata]
allow-direct-references = true
